+++ tests/repositories/test_legacy_repository.py
@@ -148,7 +148,7 @@ def test_missing_version() -> None:
 def test_get_package_information_fallback_read_setup() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("jupyter"), Version.parse("1.0.0"))
+    package = repo.package("jupyter", Version.parse("1.0.0"))
 
     assert package.source_type == "legacy"
     assert package.source_reference == repo.name
@@ -164,9 +164,7 @@ def test_get_package_information_fallback_read_setup() -> None:
 def test_get_package_information_skips_dependencies_with_invalid_constraints() -> None:
     repo = MockRepository()
 
-    package = repo.package(
-        canonicalize_name("python-language-server"), Version.parse("0.21.2")
-    )
+    package = repo.package("python-language-server", Version.parse("0.21.2"))
 
     assert package.name == "python-language-server"
     assert package.version.text == "0.21.2"
@@ -197,6 +195,15 @@ def test_get_package_information_skips_dependencies_with_invalid_constraints() -
         Dependency("pyflakes", ">=1.6.0"),
         Dependency("yapf", "*"),
     ]
+
+
+def test_package_not_canonicalized() -> None:
+    repo = MockRepository()
+
+    package = repo.package("discord.py", Version.parse("2.0.0"))
+
+    assert package.name == "discord-py"
+    assert package.pretty_name == "discord.py"
 
 
 def test_find_packages_no_prereleases() -> None:
@@ -255,7 +262,7 @@ def test_find_packages_yanked(constraint: str, expected: list[str]) -> None:
 def test_get_package_information_chooses_correct_distribution() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("isort"), Version.parse("4.3.4"))
+    package = repo.package("isort", Version.parse("4.3.4"))
 
     assert package.name == "isort"
     assert package.version.text == "4.3.4"
@@ -268,7 +275,7 @@ def test_get_package_information_chooses_correct_distribution() -> None:
 def test_get_package_information_includes_python_requires() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("futures"), Version.parse("3.2.0"))
+    package = repo.package("futures", Version.parse("3.2.0"))
 
     assert package.name == "futures"
     assert package.version.text == "3.2.0"
@@ -280,7 +287,7 @@ def test_get_package_information_sets_appropriate_python_versions_if_wheels_only
 ):
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("futures"), Version.parse("3.2.0"))
+    package = repo.package("futures", Version.parse("3.2.0"))
 
     assert package.name == "futures"
     assert package.version.text == "3.2.0"
@@ -290,7 +297,7 @@ def test_get_package_information_sets_appropriate_python_versions_if_wheels_only
 def test_get_package_from_both_py2_and_py3_specific_wheels() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("ipython"), Version.parse("5.7.0"))
+    package = repo.package("ipython", Version.parse("5.7.0"))
 
     assert package.name == "ipython"
     assert package.version.text == "5.7.0"
@@ -328,9 +335,7 @@ def test_get_package_from_both_py2_and_py3_specific_wheels() -> None:
 def test_get_package_from_both_py2_and_py3_specific_wheels_python_constraint() -> None:
     repo = MockRepository()
 
-    package = repo.package(
-        canonicalize_name("poetry-test-py2-py3-metadata-merge"), Version.parse("0.1.0")
-    )
+    package = repo.package("poetry-test-py2-py3-metadata-merge", Version.parse("0.1.0"))
 
     assert package.name == "poetry-test-py2-py3-metadata-merge"
     assert package.version.text == "0.1.0"
@@ -340,7 +345,7 @@ def test_get_package_from_both_py2_and_py3_specific_wheels_python_constraint() -
 def test_get_package_with_dist_and_universal_py3_wheel() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("ipython"), Version.parse("7.5.0"))
+    package = repo.package("ipython", Version.parse("7.5.0"))
 
     assert package.name == "ipython"
     assert package.version.text == "7.5.0"
@@ -368,7 +373,7 @@ def test_get_package_with_dist_and_universal_py3_wheel() -> None:
 def test_get_package_retrieves_non_sha256_hashes() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("ipython"), Version.parse("7.5.0"))
+    package = repo.package("ipython", Version.parse("7.5.0"))
 
     expected = [
         {
@@ -387,7 +392,7 @@ def test_get_package_retrieves_non_sha256_hashes() -> None:
 def test_get_package_retrieves_non_sha256_hashes_mismatching_known_hash() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("ipython"), Version.parse("5.7.0"))
+    package = repo.package("ipython", Version.parse("5.7.0"))
 
     expected = [
         {
@@ -410,7 +415,7 @@ def test_get_package_retrieves_non_sha256_hashes_mismatching_known_hash() -> Non
 def test_get_package_retrieves_packages_with_no_hashes() -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name("jupyter"), Version.parse("1.0.0"))
+    package = repo.package("jupyter", Version.parse("1.0.0"))
 
     assert [
         {
@@ -432,7 +437,7 @@ def test_package_yanked(
 ) -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name(package_name), Version.parse(version))
+    package = repo.package(package_name, Version.parse(version))
 
     assert package.name == package_name
     assert str(package.version) == version
@@ -446,11 +451,11 @@ def test_package_partial_yank():
             return super()._get_page(f"/{endpoint.strip('/')}_partial_yank/")
 
     repo = MockRepository()
-    package = repo.package(canonicalize_name("futures"), Version.parse("3.2.0"))
+    package = repo.package("futures", Version.parse("3.2.0"))
     assert len(package.files) == 2
 
     repo = SpecialMockRepository()
-    package = repo.package(canonicalize_name("futures"), Version.parse("3.2.0"))
+    package = repo.package("futures", Version.parse("3.2.0"))
     assert len(package.files) == 1
     assert package.files[0]["file"].endswith(".tar.gz")
 
@@ -467,7 +472,7 @@ def test_find_links_for_package_yanked(
 ) -> None:
     repo = MockRepository()
 
-    package = repo.package(canonicalize_name(package_name), Version.parse(version))
+    package = repo.package(package_name, Version.parse(version))
     links = repo.find_links_for_package(package)
 
     assert len(links) == 1